backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
dulwich>=0.21.0
paramiko>=3.0.0
pyjwt[crypto]>=2.8.0
requests>=2.31.0
numpy>=1.24.0
sympy>=1.12
pytest
httpx
